Major rewrite of PayloadReceiveChannel.
The previous version of PayloadReceiveChannel used a round-robin timesharing
approach to march through the various states.  This resulted in things
eventually falling into a state where every channel in a PayloadInputEngine
was in lockstep, each one delivering a single payload at a time, which caused
the more prolific channels to back up and eventually the run fell over.  (This
analysis, by the way, was mostly done by Kael and John J., with a tiny amount
of help from me.)

The rewritten PayloadReceiveChannel sucks in all available payloads at once,
so that the noisy channels are quickly drained.  This appears to have fixed a
problem which had dogged the DAQ for over a year!

Note that this rewrite is still an uneasy mixture of the old lockstep state
machine and my more freeform approach, and should definitely be looked at
again under less high-pressure circumstances.